requirements.txt
flask==2.3.3
flask-cors==4.0.0
groq==0.4.2
Pillow==10.4.0
python-dotenv==1.0.0
opencv-python==4.9.0.80
numpy==1.24.3
werkzeug==2.3.7
requests==2.31.0